🐛 fix: move isUpdateSnoozed to separate module to satisfy react-refresh lint rule
The react-refresh/only-export-components rule requires that a file
only exports React components. WhatsNewModal.tsx was also exporting
the non-component function isUpdateSnoozed, causing CI lint failure.

Move isUpdateSnoozed and snoozeUpdate into a dedicated
whatsNew/snooze.ts module and update all importers.
