+++ web/src/components/layout/navbar/UpdateIndicator.tsx
@@ -5,7 +5,8 @@ import { useVersionCheck } from '../../../hooks/useVersionCheck'
 import { useUpgradeState } from '../../../hooks/useUpgradeState'
 import { useFeatureHints } from '../../../hooks/useFeatureHints'
 import { FeatureHintTooltip } from '../../ui/FeatureHintTooltip'
-import { WhatsNewModal, isUpdateSnoozed } from '../../updates/WhatsNewModal'
+import { WhatsNewModal } from '../../updates/WhatsNewModal'
+import { isUpdateSnoozed } from '../../updates/whatsNew/snooze'
 import { useModalState } from '../../../lib/modals'
 import { isDemoMode } from '../../../lib/demoMode'
 import { useToast } from '../../ui/Toast'
